update producers dashboard and create project pages
